--- endpoints
@@ -1,34 +1,0 @@
-Endpoints
-
-person
-
-driver
-    delete
-    put
-    post
-    get
-    driver/driver_id/passengers         // retrieves drivers current passengers
-    driver/driver_id/accept/rider_id    // only valid if match already exists
-
-
-rider
-    delete
-    put
-    post
-    get
-    rider/rider_id/matches
-    rider/rider_id/request/driver_id
-
-
-match?
-    get
-    match/rider_id, driver_id/true
-
-----------------------------------
-
-find_matches(driver, rider)
-    checks conditionals
-    directions api
-    for each rider returns list of matches
-    
-
