+++ test/sql/cera.sql
