+++ test/ts/MatchTest.ts
@@ -1,10 +1,10 @@
-/// <reference path="../typings/mocha/mocha.d.ts" />
-/// <reference path="../typings/chai/chai.d.ts" />
+/// <reference path="../../typings/mocha/mocha.d.ts" />
+/// <reference path="../../typings/chai/chai.d.ts" />
 
 import {expect} from 'chai';
 import {map_riders_to_drivers} from "../match";
 import {Rider, Driver, Coordinate, Match, Time} from "../class_defs";
-
+import * as blank from "../random";
 
 describe('Matching', () => {
